refactore - Operating on refactoring a promo section component
