+++ src/sections/promo/style.css
@@ -9,20 +9,37 @@
   background: var(--gradient), url('../../assets/svg/common-bg.svg');
 }
 
-.promo__title {
-  font-size: 3rem;
+.promo__content{
+  max-width: 80rem;
+  margin: 0 auto;
+}
+
+.promo__greeting {
+  margin-bottom: 2rem;
+  font-size: 8rem;
   font-weight: 700;
-  letter-spacing: 2px;
+  text-transform: uppercase;
+  letter-spacing: 5px;
+  /* text-align: center; */
+  color: var(--white);
+}
+
+/* .promo__greeting span {
+  display: block;
+} */
+
+.promo__slogan {
+  margin-bottom: 2rem;
+  font-size: 18px;
   text-align: center;
+  line-height: 24px;
   color: var(--white);
 }
 
 .promo__text {
-  max-width: 80rem;
-  margin: 3rem auto 0 auto;
-  font-size: 2.2rem;
+  font-size: 18px;
   text-align: center;
-  line-height: 1.6;
+  line-height: 24px;
   color: var(--white);
 }
