+++ src/utils/i18n/locales/ru/ru.json
@@ -50,7 +50,8 @@
     "projects": "Проекты"
   },
   "promo": {
-    "promoTitle": "Привет, народ! Я Жилин Антон - разработчик!",
+    "promoGreeting": "Привет! Я Жилин Антон!",
+    "promoSlogan": "Фронтенд-разработчик из Москвы, увлечённый созданием доступных и удобных для пользователей сайтов.",
     "promoText": "Добро пожаловать в мир, где пиксели оживают, а цифровые мечты становятся реальностью. Здесь я создаю бесшовные цифровые впечатления через код — элегантные, запоминающиеся и долговечные.",
     "promoBtn": "Скачать резюме"
   },
