+++ src/utils/spliteStringRegExp.js
